Bug fix and Knockout enhancement
The method I used for using the current time on _all_ new objects was
actually using the time that the application started - that's fixed. I
also converted a bunch of nasty jQuery code to the Knockout model. This
allowed me to keep HTML/template code in the html files and let Knockout
work its magic instead. Much more efficient. :-)